Struct google_storagetransfer1::ObjectConditions

pub struct ObjectConditions {
    pub max_time_elapsed_since_last_modification: Option<String>,
    pub include_prefixes: Option<Vec<String>>,
    pub exclude_prefixes: Option<Vec<String>>,
    pub min_time_elapsed_since_last_modification: Option<String>,
}

Conditions that determine which objects will be transferred.

This type is not used in any activity, and only used as part of another schema.

Fields

max_time_elapsed_since_last_modification

maxTimeElapsedSinceLastModification is the complement to minTimeElapsedSinceLastModification.

include_prefixes

If includePrefixes is specified, objects that satisfy the object conditions must have names that start with one of the includePrefixes and that do not start with any of the excludePrefixes. If includePrefixes is not specified, all objects except those that have names starting with one of the excludePrefixes must satisfy the object conditions. Requirements: * Each include-prefix and exclude-prefix can contain any sequence of Unicode characters, of max length 1024 bytes when UTF8-encoded, and must not contain Carriage Return or Line Feed characters. Wildcard matching and regular expression matching are not supported. * None of the include-prefix or the exclude-prefix values can be empty, if specified. * Each include-prefix must include a distinct portion of the object namespace, i.e., no include-prefix may be a prefix of another include-prefix. * Each exclude-prefix must exclude a distinct portion of the object namespace, i.e., no exclude-prefix may be a prefix of another exclude-prefix. * If includePrefixes is specified, then each exclude-prefix must start with the value of a path explicitly included by includePrefixes. The max size of includePrefixes is 20.

exclude_prefixes

excludePrefixes must follow the requirements described for includePrefixes. The max size of excludePrefixes is 20.

min_time_elapsed_since_last_modification

If unspecified, minTimeElapsedSinceLastModification takes a zero value and maxTimeElapsedSinceLastModification takes the maximum possible value of Duration. Objects that satisfy the object conditions must either have a lastModificationTime greater or equal to NOW - maxTimeElapsedSinceLastModification and less than NOW - minTimeElapsedSinceLastModification, or not have a lastModificationTime.