Struct google_storagetransfer1::Operation

pub struct Operation {
    pub metadata: Option<HashMap<String, String>>,
    pub done: Option<bool>,
    pub response: Option<HashMap<String, String>>,
    pub name: Option<String>,
    pub error: Option<Status>,
}

This resource represents a long-running operation that is the result of a network API call.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get transfer operations (response)

Fields

metadata

Represents the transfer operation object.

done

If the value is false, it means the operation is still in progress. If true, the operation is completed and the result is available.

response

The normal response of the operation in case of success. If the original method returns no data on success, such as Delete, the response is google.protobuf.Empty. If the original method is standard Get/Create/Update, the response should be the resource. For other methods, the response should have the type XxxResponse, where Xxx is the original method name. For example, if the original method name is TakeSnapshot(), the inferred response type is TakeSnapshotResponse.

name

The server-assigned name, which is only unique within the same service that originally returns it. If you use the default HTTP mapping above, the name should have the format of operations/some/unique/name.

error

The error result of the operation in case of failure.