Struct google_storagetransfer1::TransferCounters

pub struct TransferCounters {
    pub bytes_from_source_skipped_by_sync: Option<String>,
    pub objects_found_from_source: Option<String>,
    pub bytes_failed_to_delete_from_sink: Option<String>,
    pub objects_failed_to_delete_from_sink: Option<String>,
    pub objects_found_only_from_sink: Option<String>,
    pub objects_from_source_skipped_by_sync: Option<String>,
    pub bytes_copied_to_sink: Option<String>,
    pub bytes_found_from_source: Option<String>,
    pub objects_deleted_from_source: Option<String>,
    pub bytes_found_only_from_sink: Option<String>,
    pub bytes_from_source_failed: Option<String>,
    pub bytes_deleted_from_source: Option<String>,
    pub bytes_deleted_from_sink: Option<String>,
    pub objects_copied_to_sink: Option<String>,
    pub objects_deleted_from_sink: Option<String>,
    pub objects_from_source_failed: Option<String>,
}

A collection of counters that report the progress of a transfer operation.

This type is not used in any activity, and only used as part of another schema.

Fields
bytes_from_source_skipped_by_sync: Option<String>

Bytes in the data source that are not transferred because they already exist in the data sink.

objects_found_from_source: Option<String>

Objects found in the data source that are scheduled to be transferred, which will be copied, excluded based on conditions, or skipped due to failures.

bytes_failed_to_delete_from_sink: Option<String>

Bytes that failed to be deleted from the data sink.

objects_failed_to_delete_from_sink: Option<String>

Objects that failed to be deleted from the data sink.

objects_found_only_from_sink: Option<String>

Objects found only in the data sink that are scheduled to be deleted.

objects_from_source_skipped_by_sync: Option<String>

Objects in the data source that are not transferred because they already exist in the data sink.

bytes_copied_to_sink: Option<String>

Bytes that are copied to the data sink.

bytes_found_from_source: Option<String>

Bytes found in the data source that are scheduled to be transferred, which will be copied, excluded based on conditions, or skipped due to failures.

objects_deleted_from_source: Option<String>

Objects that are deleted from the data source.

bytes_found_only_from_sink: Option<String>

Bytes found only in the data sink that are scheduled to be deleted.

bytes_from_source_failed: Option<String>

Bytes in the data source that failed during the transfer.

bytes_deleted_from_source: Option<String>

Bytes that are deleted from the data source.

bytes_deleted_from_sink: Option<String>

Bytes that are deleted from the data sink.

objects_copied_to_sink: Option<String>

Objects that are copied to the data sink.

objects_deleted_from_sink: Option<String>

Objects that are deleted from the data sink.

objects_from_source_failed: Option<String>

Objects in the data source that failed during the transfer.