Struct google_storagetransfer1::TransferJob

pub struct TransferJob {
    pub transfer_spec: Option<TransferSpec>,
    pub status: Option<String>,
    pub deletion_time: Option<String>,
    pub description: Option<String>,
    pub schedule: Option<Schedule>,
    pub project_id: Option<String>,
    pub last_modification_time: Option<String>,
    pub creation_time: Option<String>,
    pub name: Option<String>,
}

This resource represents the configuration of a transfer job that runs periodically.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list transfer jobs (none)

create transfer jobs (request|response)

get transfer jobs (response)

patch transfer jobs (response)

Fields

transfer_spec

Transfer specification. Required.

status

Status of the job. This value MUST be specified for CreateTransferJobRequests. NOTE: The effect of the new job status takes place during a subsequent job run. For example, if you change the job status from ENABLED to DISABLED, and an operation spawned by the transfer is running, the status change would not affect the current operation.

deletion_time

This field cannot be changed by user requests.

description

A description provided by the user for the job. Its max length is 1024 bytes when Unicode-encoded.

schedule

Schedule specification. Required.

project_id

The ID of the Google Developers Console project that owns the job. Required.

last_modification_time

This field cannot be changed by user requests.

creation_time

This field cannot be changed by user requests.

name

A globally unique name assigned by Storage Transfer Service when the job is created. This field should be left empty in requests to create a new transfer job; otherwise, the requests result in an INVALID_ARGUMENT error.