Struct google_storagetransfer1::TransferJobMethods

pub struct TransferJobMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on transferJob resources. It is not used directly, but through the Storagetransfer hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_storagetransfer1 as storagetransfer1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use storagetransfer1::Storagetransfer;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Storagetransfer::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `create(...)`, `get(...)`, `list(...)` and `patch(...)`
// to build up your call.
let rb = hub.transfer_jobs();

Methods

impl<'a, C, A> TransferJobMethods<'a, C, A>

fn list(&self) -> TransferJobListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists transfer jobs.

fn patch(&self, request: UpdateTransferJobRequest, job_name: &str) -> TransferJobPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a transfer job. Updating a job's transfer spec does not affect transfer operations that are running already. Updating the scheduling of a job is not allowed.

Arguments

request - No description provided.

jobName - The name of job to update. Required.

fn create(&self, request: TransferJob) -> TransferJobCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a transfer job that runs periodically.

Arguments

request - No description provided.

fn get(&self, job_name: &str) -> TransferJobGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a transfer job.

Arguments

jobName - The job to get. Required.