Struct google_storagetransfer1::TransferSpec

pub struct TransferSpec {
    pub http_data_source: Option<HttpData>,
    pub transfer_options: Option<TransferOptions>,
    pub object_conditions: Option<ObjectConditions>,
    pub gcs_data_source: Option<GcsData>,
    pub gcs_data_sink: Option<GcsData>,
    pub aws_s3_data_source: Option<AwsS3Data>,
}

Configuration for running a transfer.

This type is not used in any activity, and only used as part of another schema.

Fields
http_data_source: Option<HttpData>

An HTTP URL data source.

transfer_options: Option<TransferOptions>

If the option deleteObjectsUniqueInSink is true, object conditions based on objects' lastModificationTime are ignored and do not exclude objects in a data source or a data sink.

object_conditions: Option<ObjectConditions>

Only objects that satisfy these object conditions are included in the set of data source and data sink objects. Object conditions based on objects' lastModificationTime do not exclude objects in a data sink.

gcs_data_source: Option<GcsData>

A Google Cloud Storage data source.

gcs_data_sink: Option<GcsData>

A Google Cloud Storage data sink.

aws_s3_data_source: Option<AwsS3Data>

An AWS S3 data source.